split test spec extensions into extension classes
