+++ test/core/Do.Test.Architecture/Extensions/IntegerExtensions.cs
@@ -1,0 +1,8 @@
+using Do.Testing;
+
+namespace Do.Test;
+
+public static class IntegerExtensions
+{
+    public static int AnInt(this Stubber _) => 42;
+}
