+++ test/core/Do.Test.Architecture/Extensions/StringExtensions.cs
@@ -1,0 +1,8 @@
+using Do.Testing;
+
+namespace Do.Test;
+
+public static class StringExtensions
+{
+    public static string AString(this Stubber _) => "test";
+}
